// frontend/src/api/gen/schemas/appointment.ts
/**
 * Generated by orval v7.1.1 🍺
 * Do not edit manually.
 * api
 * OpenAPI spec version: 0.1.0
 */

export interface Appointment {
  /** @nullable */
  content?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  end?: string | null;
  event_id: string;
  id: string;
  start: string;
  title: string;
}