// frontend/src/api/gen/schemas/appointmentForUpdate.ts
/**
 * Generated by orval v7.1.1 🍺
 * Do not edit manually.
 * api
 * OpenAPI spec version: 0.1.0
 */

export interface AppointmentForUpdate {
  /** @nullable */
  content?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  end?: string | null;
  /** @nullable */
  start?: string | null;
  /** @nullable */
  title?: string | null;
}